move classes into app folder
